refactor: remove additional selection variables, put constants in separate file
