+++ webapp/javascript/pages/Heatmap/Heatmap.module.scss
@@ -5,6 +5,7 @@ $svg-side-margin: 40px;
   width: calc(100% - $svg-side-margin * 2);
   margin: auto auto $svg-side-margin auto;
   min-height: 100px;
+  user-select: none;
 
   span[role='progressbar'] {
     display: block;
